Convert the distilled HTML into JSON, when the conversion is defined (#18)
This is only supported in the JavaScript version for now. To give it a
try, launch as usual with `./middleman.js` and then open
`localhost:3000`. Pick either the example for NYTimes Best-sellers,
Slashdot discussion, or Goodreads (other examples do not have any
conversion defined yet). Expect to see data in JSON as the very final
step.

To check on the CLI, run something like:

```bash
./middleman.js distill https://www.nytimes.com/books/best-sellers/
```

which should print out the best-sellers list in pretty JSON.

The conversion itself is defined by specifying the selector for
iterating the row, and for each column/field. This is very similar with
the so-calling locator parsing in MCP GetGather.