Support for radio buttons, demonstrated on Best Buy
To verify, launch `middleman.js` or `middleman.py` and then open:
`localhost:3000/start?location=bestbuy.com/profile/ss/orderlookup`.

This should start Best Buy sign-in process, which start with the email
input and after that, there is a set of radio buttons to choose the delivery
methods for the verification code.

The distilled page correctly matches some of the radio buttons (intentionally
not all of them, to illustrate the concept). When one of them is chosen,
pressing the submit button will also correctly choose the same radio button
on the upstream Best Buy page.

Note that the Best Buy patterns are not completed, they are not designed to
correctly extract the order history but merely to check this radio button
functionality.

To test manually on the CLI, run:

```bash
./middleman.py run bestbuy.com/profile/ss/orderlookup
```