Properly handle checkboxes (#39)
Typically, a checkbox is used to denote "Remember me", to persist
the login session. Rather than using the autoclick hack, this change
handles it by properly relaying the checked status from the distilled
page to the actual web page.

Verifying is tricky. The best way is to run the ACME test server from
the MCP GetGather project and make sure it's running on localhost:5001.

Then run `./middleman.js` or `./middleman.py` as usual, and open this
in the browser:

http://localhost:3000/start?location=http://localhost:5001/auth/email-and-password-checkbox

This will initiate the distillation loop on the ACME test page which
contains the email and password and also the "Remember me" checkbox.
The checkbox on that ACME test page will be checked, as the automation
progresses.