Add a safeguard against null element during JSON conversion (#41)
This is applying the same safeguard which is already in the Python code,
to the JavaScript version.

To verify, run the distillation loop for NY Times Best-sellers:

```bash
./middleman.js run www.nytimes.com/books/best-sellers
```

Before this change, the conversion from HTML to JSON will fail.

With this change, the final JSON result is obtained without any issue.