Prioritize checking for termination in the distillation loop (#40)
This is to prepare for the separation of autoclick and form submission.

To simplify the logic, the first step in the distillation loop should be
checking for a possible termination (gg-stop). As a bonus, this will give
some extra time for the upstream website to process any form submission
(carried out typically as the last step in the loop).

To verify, run in the CLI as usual, e.g.

```bash
./middleman.py run bbc.com/saved
```

Also, run the web UI with `./middleman.py` and opening `localhost:3000`.
Then verify with NY Times Best-sellers, NBA key dates, etc etc.