Tweak the separation between autoclick and form submission (#45)
This is just a minor tweak to PR #38, with the goal of making it
semantically more explicit that there are two distinctive steps:

* clicking on any elements with gg-autoclick attribute
* submitting the form

The second step is only carried out when all the form fields are
properly filled.

To verify, just launch and test the usual examples (Goodreads, Amazon,
NYTimes Best-sellers, NBA key dates, etc).