Migrate to Zendriver
Instead of using Patchright, which is also locked to a certain working
version of Chromium, this is based on Zendriver which uses the system
Chromium. Unfortunately, since it's Python only, we have to lose the
JavaScript version of Middleman.

Also, pseudo-class like `:has-text()` is not supported anymore, since
this is a specific Playwright feature. There is no exact equivalent,
however since XPath selector can be used, the approximate variant is
e.g. `//span[contains(text(), ...]`.

Note that only some of the distillation patterns have been converted
from the pseudo-class `:has-text()`.

Also, iframes are pretty much not handled for now.