Support HTTP(s) proxy with BROWSER_HTTP_PROXY env (#62)
Zendriver (due to Chromium/CDP) does not support setting the proxy
username and password in the launch argument. To facilitate the use
of authenticated proxy, we ought to intercept the callback and supply
the credentials manually.

To verify, set `BROWSER_HTTP_PROXY` to point to a proxy server which
requires authentication. Launch Middleman as usual, and every thing
should still work as expected.